Revert "lwnbd: improve opl integration"
