+++ download_lwNBD.sh
@@ -3,7 +3,7 @@
 ## Download lwNBD
 REPO_URL="https://github.com/bignaux/lwNBD.git"
 REPO_FOLDER="modules/network/lwNBD"
-COMMIT="0fed27b81afb8e7619d000599b17a8c272b14dd5"
+COMMIT="9777a10f840679ef89b1ec6a588e2d93803d7c37"
 if test ! -d "$REPO_FOLDER"; then
   git clone $REPO_URL "$REPO_FOLDER" || { exit 1; }
   (cd $REPO_FOLDER && git checkout "$COMMIT" && cd -) || { exit 1; }
