+++ include/opl.h
@@ -76,6 +76,7 @@ int saveConfig(int types, int showUI);
 void applyConfig(int themeID, int langID);
 void menuDeferredUpdate(void *data);
 void moduleUpdateMenu(int mode, int themeChanged, int langChanged);
+void handleLwnbdSrv();
 void deinit(int exception, int modeSelected);
 
 // Shutdown minimal services initiated for auto loading.
